+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog of lizard-nxt client
 Unreleased (1.2.7) (XXXX-XX-XX)
 -------------------------------
 
+- Fix bug with bar size when event.
+
+- Throw error when no backend is up and running.
+
 - Add credentials to UTFGrid requests.
 
 - Add domains for sandbox rewrites.